// src/enums/enum.ts
export enum ENUM_USER_ROLE {
  SUPER_ADMIN = 'super-admin',
  ADMIN = 'admin',
  GUEST = 'guest',
  HOST = 'host'
}

export enum ENUM_USER_STATUS {
  ACTIVE = 'active',
  BLOCKED = 'blocked',
}

export enum ENUM_bOOKING_STATUS {
 ACTIVE = 'active',
 BOOKED = 'booked',
 COMPLETED = 'completed',
 CANCELLED = 'cancelled'
}

export enum ENUM_BOOKING_TYPE {
  DIRECT = 'direct',
  RESERVE = 'reserve'
}

export enum ENUM_PAYMENT_STATUS {
  PAID = 'paid',
  UNPAID = 'unpaid'
}